refactor(core): move CrawlerResult to core/models.py; type CrawlerInfo.query
CrawlerResult lived in crawlers/__init__.py, preventing CrawlerInfo.query
(defined in core/models.py) from being typed to return it without a circular
import. Moving CrawlerResult to core breaks the cycle cleanly.

discover_crawlers() now wraps each raw dict-returning query function via
_wrap_query so all CrawlerInfo.query callsites receive a typed CrawlerResult.
runner.py updated to use attribute access (result.solved etc.).

CrawlerResult is re-exported from both ojhunt.core and ojhunt.crawlers —
both import paths remain valid. QueryResult and NullCrawler stay in core
as pipeline/runner concepts, not crawler concepts.
